refactor: moved EmployeeAuthority logic to employee
