+++ src/main/java/com/pawelnu/projectmanager/config/DataInit.java
@@ -3,14 +3,14 @@ package com.pawelnu.projectmanager.config;
 import com.github.javafaker.Faker;
 import com.pawelnu.projectmanager.endpoints.authority.AuthorityEntity;
 import com.pawelnu.projectmanager.endpoints.authority.AuthorityRepository;
-import com.pawelnu.projectmanager.endpoints.authority.employee.EmployeeAuthorityEntity;
-import com.pawelnu.projectmanager.endpoints.authority.employee.EmployeeAuthorityRepository;
 import com.pawelnu.projectmanager.endpoints.company.CompanyEntity;
 import com.pawelnu.projectmanager.endpoints.company.CompanyRepository;
 import com.pawelnu.projectmanager.endpoints.company.address.CompanyAddressEntity;
 import com.pawelnu.projectmanager.endpoints.company.address.CompanyAddressRepository;
 import com.pawelnu.projectmanager.endpoints.company.employee.EmployeeEntity;
 import com.pawelnu.projectmanager.endpoints.company.employee.EmployeeRepository;
+import com.pawelnu.projectmanager.endpoints.company.employee.authority.EmployeeAuthorityEntity;
+import com.pawelnu.projectmanager.endpoints.company.employee.authority.EmployeeAuthorityRepository;
 import jakarta.annotation.PostConstruct;
 import java.util.ArrayList;
 import java.util.List;
