+++ src/main/java/com/pawelnu/projectmanager/endpoints/authority/AuthorityApi.java
@@ -101,20 +101,4 @@ public interface AuthorityApi {
   @DeleteMapping("/{id}")
   @Operation(description = "Delete authority by id.")
   ResponseEntity<SimpleResponse> deleteById(@Parameter() @PathVariable() UUID id);
-
-  @ApiResponse(
-      responseCode = "201",
-      description = "Created",
-      content = {
-        @Content(
-            mediaType = MediaType.APPLICATION_JSON_VALUE,
-            schema = @Schema(implementation = AddAuthorityToUserResponseDTO.class))
-      })
-  @ResponseErrors
-  @PostMapping("/add-authority-to-employee")
-  @Operation(description = "Add authority to user.")
-  ResponseEntity<AddAuthorityToUserResponseDTO> addAuthorityToEmployee(
-      @Valid @RequestBody AddAuthorityToUserRequestDTO body);
-  //  TODO add delete authority from user
-  //  TODO add test delete authority from user
 }
