+++ src/main/java/com/pawelnu/projectmanager/endpoints/authority/AuthorityEntity.java
@@ -1,6 +1,6 @@
 package com.pawelnu.projectmanager.endpoints.authority;
 
-import com.pawelnu.projectmanager.endpoints.authority.employee.EmployeeAuthorityEntity;
+import com.pawelnu.projectmanager.endpoints.company.employee.authority.EmployeeAuthorityEntity;
 import com.pawelnu.projectmanager.entity.Auditable;
 import jakarta.persistence.Entity;
 import jakarta.persistence.GeneratedValue;
