+++ src/main/java/com/pawelnu/projectmanager/endpoints/company/employee/authority/EmployeeAuthorityMapper.java
@@ -1,4 +1,4 @@
-package com.pawelnu.projectmanager.endpoints.authority.employee;
+package com.pawelnu.projectmanager.endpoints.company.employee.authority;
 
 import com.pawelnu.projectmanager.endpoints.authority.AddAuthorityToUserResponseDTO;
 import org.mapstruct.Mapper;
